rename _constants to constants and reference them from tests instead of redefining the constants
