+++ algopy_testing/tests/primitives/test_uint64.py
@@ -5,11 +5,9 @@ from collections.abc import Callable
 import algokit_utils
 import pytest
 from algopy import UInt64
+from algopy.constants import MAX_UINT64, MAX_UINT512
 
 from tests.primitives.conftest import AVMInvoker
-
-MAX_UINT64 = 2**64 - 1
-MAX_UINT512 = 2**512 - 1
 
 _negative_value_error = "expected positive value"
 _too_big64_error = re.escape(f"expected value <= {MAX_UINT64}")
